# service/apigateway/docs/swagger.yaml
definitions:
...
  requests.CreateMerchantDocumentRequest:
    properties:
      document_type:
        type: string
      document_url:
        type: string
      merchant_id:
        minimum: 1
        type: integer
    required:
    - document_type
    - document_url
    - merchant_id
    type: object
...
  response.ApiResponseMerchantDocument:
    properties:
      data:
        $ref: '#/definitions/response.MerchantDocumentResponse'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseMerchantDocumentAll:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseMerchantDocumentDelete:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ApiResponsePaginationMerchantDocument:
    properties:
      data:
        items:
          $ref: '#/definitions/response.MerchantDocumentResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationMerchantDocumentDeleteAt:
    properties:
      data:
        items:
          $ref: '#/definitions/response.MerchantDocumentResponseDeleteAt'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/merchant-documents:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all merchant documents
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of merchant documents
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantDocument'
        "500":
          description: Failed to retrieve merchant document data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all merchant documents
      tags:
      - Merchant Document
  /api/merchant-documents/{id}:
    get:
      consumes:
      - application/json
      description: Get a merchant document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Document details
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to get document
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get merchant document by ID
      tags:
      - Merchant Document
  /api/merchant-documents/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all active merchant documents
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active merchant documents
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantDocument'
        "500":
          description: Failed to retrieve active merchant documents
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all active merchant documents
      tags:
      - Merchant Document
  /api/merchant-documents/create:
    post:
      consumes:
      - application/json
      description: Create a new document for a merchant
      parameters:
      - description: Create merchant document request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.CreateMerchantDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created document
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Bad request or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to create document
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new merchant document
      tags:
      - Merchant Document
  /api/merchant-documents/permanent/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a merchant document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted document
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocumentDelete'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to delete document
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a merchant document
      tags:
      - Merchant Document
  /api/merchant-documents/permanent/all:
    post:
      consumes:
      - application/json
      description: Permanently delete all merchant documents from the database
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted all documents permanently
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocumentAll'
        "500":
          description: Failed to permanently delete all documents
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete all merchant documents
      tags:
      - Merchant Document
  /api/merchant-documents/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore a merchant document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Restored document
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Bad request or invalid ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore document
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a merchant document
      tags:
      - Merchant Document
  /api/merchant-documents/restore/all:
    post:
      consumes:
      - application/json
      description: Restore all merchant documents that were previously deleted
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored all documents
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocumentAll'
        "500":
          description: Failed to restore all documents
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore all merchant documents
      tags:
      - Merchant Document
  /api/merchant-documents/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all trashed merchant documents
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed merchant documents
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantDocumentDeleteAt'
        "500":
          description: Failed to retrieve trashed merchant documents
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all trashed merchant documents
      tags:
      - Merchant Document
  /api/merchant-documents/trashed/{id}:
    post:
      consumes:
      - application/json
      description: Trashed a merchant document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trashed document
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Bad request or invalid ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to trashed document
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Trashed a merchant document
      tags:
      - Merchant Document
  /api/merchant-documents/update/{id}:
    post:
      consumes:
      - application/json
      description: Update a merchant document with the given ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update merchant document request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateMerchantDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated document
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Bad request or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to update document
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update a merchant document
      tags:
      - Merchant Document
  /api/merchants:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all merchants
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of merchants
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchant'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all merchants
      tags:
      - Merchant
  /api/merchants-documents/update-status/{id}:
    post:
      consumes:
      - application/json
      description: Update the status of a merchant document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update status request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateMerchantDocumentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated document
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Bad request or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to update document status
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update merchant document status
      tags:
      - Merchant Document
  /api/merchants/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a merchant by its ID permanently
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted merchant
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDelete'
        "400":
          description: Bad request or invalid ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to delete merchant
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a merchant permanently
      tags:
      - Merchant
    get:
      consumes:
      - application/json
      description: Retrieve a merchant by its ID.
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Merchant data
          schema:
            $ref: '#/definitions/response.ApiResponseMerchant'
        "400":
          description: Invalid merchant ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find a merchant by ID
      tags:
      - Merchant
  /api/merchants/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active merchants
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active merchants
          schema:
            $ref: '#/definitions/response.ApiResponsesMerchant'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find active merchants
      tags:
      - Merchant
...
  /api/merchants/create:
    post:
      consumes:
      - application/json
      description: Create a new merchant with the given name and user ID
      parameters:
      - description: Create merchant request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.CreateMerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created merchant
          schema:
            $ref: '#/definitions/response.ApiResponseMerchant'
        "400":
          description: Bad request or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to create merchant
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new merchant
      tags:
      - Merchant
...
  /api/merchants/restore/all:
    post:
      consumes:
      - application/json
      description: Restore all merchant records that were previously deleted.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored all merchant records
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantAll'
        "500":
          description: Failed to restore all merchant records
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore all merchant records
      tags:
      - Merchant
...
swagger: "2.0"